Zip Java
I added the json to connect my server for my websites. Then I made three functions that contact the server to get the location of what ever zip the user entered. Then I started making code that verifies/checks to see if the zip is a valid amount of numbers.